Move VR stuff to ThreeDView
