+++ demo/client/src/view.ts
@@ -10,7 +10,6 @@
 
 import * as glmatrix from 'gl-matrix';
 
-import {FAR_CLIP_PLANE, NEAR_CLIP_PLANE} from './3d-demo';
 import {AntialiasingStrategy, AntialiasingStrategyName, NoAAStrategy} from "./aa-strategy";
 import {StemDarkeningMode, SubpixelAAType} from "./aa-strategy";
 import {AAOptions} from './app-controller';
@@ -175,9 +174,6 @@ export abstract class DemoView extends PathfinderView implements RenderContext {
     protected colorBufferHalfFloatExt: any;
 
     private wantsScreenshot: boolean;
-    private vrDisplay: VRDisplay | null;
-    private vrFrameData: VRFrameData | null;
-    private inVRRAF: boolean;
 
     /// NB: All subclasses are responsible for creating a renderer in their constructors.
     constructor(areaLUT: HTMLImageElement,
@@ -199,60 +195,6 @@ export abstract class DemoView extends PathfinderView implements RenderContext {
 
         this.wantsScreenshot = false;
 
-
-        this.inVRRAF = false;
-        this.vrDisplay = null;
-        if ("VRFrameData" in window) {
-           this.vrFrameData = new VRFrameData;
-        } else {
-            this.vrFrameData = null;
-        }
-
-        this.vrSetup();
-    }
-
-    vrSetup(): void {
-        if (navigator.getVRDisplays) {
-            navigator.getVRDisplays().then((displays) => {
-              if (displays.length > 0) {
-                this.vrDisplay = displays[displays.length - 1];
-
-                // It's heighly reccommended that you set the near and far planes to
-                // something appropriate for your scene so the projection matricies
-                // WebVR produces have a well scaled depth buffer.
-                this.vrDisplay.depthNear = NEAR_CLIP_PLANE;
-                this.vrDisplay.depthFar = FAR_CLIP_PLANE;
-                unwrapNull(document.getElementById('pf-vr')).style.display = "initial";
-              } else {
-                // no vr displays
-              }
-            });
-        } else {
-            // no vr support
-        }
-
-        window.addEventListener('vrdisplaypresentchange', () => {
-          if (this.vrDisplay == null)
-              return;
-
-          if (this.vrDisplay.isPresenting) {
-            const that = this;
-            this.resized();
-            function vrCallback(): void {
-                if (that.vrDisplay == null || !that.renderer.inVR) {
-                    return;
-                }
-                that.vrDisplay.requestAnimationFrame(vrCallback);
-                that.inVRRAF = true;
-                that.redraw();
-                that.inVRRAF = false;
-            }
-            this.vrDisplay.requestAnimationFrame(vrCallback);
-          } else {
-            this.renderer.inVR = false;
-            this.resized();
-          }
-        });
     }
 
     attachMeshes(meshes: PathfinderPackedMeshes[]): void {
@@ -304,11 +246,9 @@ export abstract class DemoView extends PathfinderView implements RenderContext {
         return buffer;
     }
 
-    enterVR(): void {
-        if (this.vrDisplay != null) {
-            this.renderer.inVR = true;
-            this.vrDisplay.requestPresent([{ source: this.canvas }]);
-        }
+    enterVR(): void {}
+    redrawVR(): void {
+        this.renderer.redraw();
     }
 
     redraw(): void {
@@ -317,17 +257,10 @@ export abstract class DemoView extends PathfinderView implements RenderContext {
         if (!this.renderer.meshesAttached)
             return;
 
-        if (!this.renderer.inVR || this.vrDisplay == null ||
-            this.vrFrameData == null) {
+        if (!this.renderer.inVR) {
             this.renderer.redraw();
         } else {
-            if (!this.inVRRAF) {
-                // redraw() called outside of vr RAF, will get drawn later
-                return;
-            }
-            this.vrDisplay.getFrameData(this.vrFrameData);
-            this.renderer.redrawVR(this.vrFrameData);
-            this.vrDisplay.submitFrame();
+            this.redrawVR();
         }
 
         // Invoke the post-render hook.
